extract core pager into dedicated file
https://coveord.atlassian.net/browse/KIT-909
